Simplify GitRepo.retrieveBook()
The method had several unused variables and seemingly pointless logic,
considering how it is used.

The method is only called in the current scenarios:

- A remote book has no corresponding local book
- When force loading a book from the repo

Hence, I see no reason to "safely" retrieve the latest version of the
book from the repo. We know that we just want the latest version from
the repo.